VIB-588: Targeted messaging (pt. 1) & rework of participant ID (temporarily breaks netsim)
* ParticipantId is now of type std::uint64_t (breaks network compatibility by structure)
* It represents a hash of the participantName (decouples ID from static config - breaks network compatibility by semantics)
* Replaces legacyEpa values along with serviceId in messaging

* Started with targeted messaging

+++ IntegrationBus/include/ib/mw/EndpointAddress.hpp
@@ -14,7 +14,7 @@ namespace mw {
  *  via ISyncAdapter::SynchronizeDT or ISyncAdapter::SynchronizeET but
  *  not both.
 */
-using ParticipantId = uint16_t;
+using ParticipantId = uint64_t;
 
 /*! \brief Participant specific identifier for its communication endpoints.
  *
